Basic ball physics working
The basic pong physics in the pong backend appear to be working.

About the UI, it does represent the backend accurately, although I still
need to work on forcing the scale to be right to make sure it actually
does display properly, and there is an issue with choppiness, but that
can be worked on later.