Improve smoothness and framerate controls
The framerate and the tick time are now handled independently of each
other which allows for the ticks to be called more often than the frames
are actually rendered which is good since it avoids the issues with slow
framerates causing logic errors.

I also improved how smooth animations are just by removing the calls to
runding the floating point values off during the ball's tick method,
which avoids the zigzagging effect.

Lastly, fixed up the ball collision logic with paddles so that angles of
over 75 degrees are now completely unachievable, as desired.