Add scoreboard, fix resizing logic
Added a proper scoreboard for the pong game, which displays the score
live as it is changed.

Also fixed a bug (which I found by adding the scoreboard) where the ball
is sent way off the screen initially since the time passed between ticks
is equal to the current time.

Finally fixed the logic for the resizing of the pong ui window in order
to keep good scaling. It involved setting listeners for the prefWidth
and prefHeight properties of the object.