Create working pong game (#14)
* Working on actual pong backend

Began work on the pong game's backend portion, which will be the driving
force behind the UI. As of now, no UI has really been touched at all,
it's all been backend stuff.

* Complete implementation of pong backend

The pong backend is in a done state, untested however and not entirely
sure if it truly is in a done state. Theoretically however it should
work.

* Basic ball physics working

The basic pong physics in the pong backend appear to be working.

About the UI, it does represent the backend accurately, although I still
need to work on forcing the scale to be right to make sure it actually
does display properly, and there is an issue with choppiness, but that
can be worked on later.

* Changed to milliseconds for calculations

Some of the movement calcualtions were being performed with doubles and
seconds which I think partially caused the choppy animations so I
switched it to milliseconds.

* Basic pong paddles working

The UI with some pong paddles and the ball are working. Still isn't as
smooth as I'd like it to be, although I have some ideas about fixing it.

* Better paddle logic working

Got the paddles working using velocities instead of gradually increasing
the x and y values while the button is being held down, which makes for
smoother movemenets as well as the possibility of handling it via
animations later on.

* Add paddle bounds

The paddle bounds are now restricted correctly so that paddles don't go
offscreen anymore.

* Small tweaks to constants and logic

Just made a couple of changes to some of the constants for the speed of
the paddle and of the ball, and also cleaned up some of the logic too.

* Improve smoothness and framerate controls

The framerate and the tick time are now handled independently of each
other which allows for the ticks to be called more often than the frames
are actually rendered which is good since it avoids the issues with slow
framerates causing logic errors.

I also improved how smooth animations are just by removing the calls to
runding the floating point values off during the ball's tick method,
which avoids the zigzagging effect.

Lastly, fixed up the ball collision logic with paddles so that angles of
over 75 degrees are now completely unachievable, as desired.

* Bit of code cleanup

Just looked through the code and cleaned up some of the problems, with
suggestions from the IntelliJ analysis tools.

* Fixed stupid mistake.

Made a stupid error by removing something from an interface without also removing the things which implement the interace. Error has been corrected.

* Fine-tuned pong specifications

Fined tuned the specifications for the pong game, including setting the
correct width and height ratio for the board as well as setting up
better speeds for things.

* Fix up class headers

Some classes were missing class headers to describe what the classes
did. I resolved this by adding them.

* Add scoreboard, fix resizing logic

Added a proper scoreboard for the pong game, which displays the score
live as it is changed.

Also fixed a bug (which I found by adding the scoreboard) where the ball
is sent way off the screen initially since the time passed between ticks
is equal to the current time.

Finally fixed the logic for the resizing of the pong ui window in order
to keep good scaling. It involved setting listeners for the prefWidth
and prefHeight properties of the object.

* Small typo fix