Complete base main menu
The basic main menu has been completed and looks quite nice for the
first iteration.

It's completely ready for more games to be added to it, as all you would
have to do is instantiate the game in the `games` array in the main menu
and a menu item for the game will be created automatically.

Still need to work on the logic for playing the game and such.