New scene logic, improve centering
Realized that part of the issue was that there was no text in the Label
when trying to calculate its width so I added default text of "0" to get
sizing working right.

I also changed the way the main menu worked a little to use a scene
method of switching to games.