Pong advanced collision detection
Added some of the proposed ideas about advanced collision detection via
tracing back the ball's position at the possible time that it would have
collided and doing all the rest of the calculations for when the ball
was there. One thing I'm not happy about is that this may have increased
the load during the `renderTick()` method.

This does however mean that the `renderTick()` method can be run less.

Because of that fact, the `renderTick()` method is run after a frame is
rendered, so it will be ready for the next frame.

Also added a listener for when the ball collided with another object.