Fixed ExecutorService to make daemon threads.
The threads created for various purposes are now daemon, which was something that Nic
mentioned for me to check.

Daemon threads exit when the main thread exits, which is what we want since otherwise the port for multiplayer
would continue to be used after the user exits the application, which could cause them a lot of confusion.