Almost working fullscreen logic
The pong game works nicely in fullscreen and so does the main menu.

The only problem comes with any sort of popups, which makes the user exit fullscreen.
We should try to eliminate these dialogs and replace them with more appropriate things, like a controlfx dialog
or a new page.

This is working towards resolving #61.