Two bug fixes
One bug that Nic found was that after playing a network game with another player, disconnecting from the party and then playing again,
there would be a bunch of nullpointers.

This had to do with two timeline objects being instantiated by accident, and was resolved by not instantiating them twice.

Also fixed a bug where even if the user accepted a game invite, it would still send a decline message ten seconds later.